+++ api/CHANGELOG.md
@@ -13,6 +13,9 @@ https://github.com/gear-tech/gear-js/pull/1467
 - Support new vouchers according to https://github.com/gear-tech/gear/pull/3606
 - Bump polkadot dependencies to `10.11.2`
 
+https://github.com/gear-tech/gear-js/pull/1472
+- Support `uploadCode` call for vouchers.
+
 ## 0.35.2
 
 _10/31/2023_